+++ src/index.test.ts
@@ -1,6 +1,11 @@
 import { runTests, TestGroup } from "@ethossoftworks/knock-on-wood"
 import { Action, createStore, applyMiddleware, combineReducers } from "redux"
-import { createEffectReducerMiddleware, combineEffectReducers, runEffect } from "./middleware"
+import {
+    createEffectReducerMiddleware,
+    combineEffectReducers,
+    runEffect,
+    combineRootEffectReducers,
+} from "./middleware"
 import {
     dispatch,
     run,
@@ -80,7 +85,7 @@ function reducer2(state: State2 = initialState2, action: TestActions): State2 {
         case "DECREMENT_NUMBER":
             return {
                 ...state,
-                ...{ testNumber: state.testNumber2 - 1 },
+                ...{ testNumber2: state.testNumber2 - 1 },
             }
         case "ECHO_RECEIVED":
             return { ...state, ...{ echoReceived2: true } }
@@ -216,6 +221,43 @@ const Tests: TestGroup<typeof testContext> = {
                 logger.count() <= 4 &&
                     logger.first().effect.type === EffectType.Interval &&
                     isEqual(logger.last().effect, cancel("decrement_interval"))
+            )
+        },
+        testCancelGroup: async ({ assert, context: { dispatched, logger, middlewareContext } }) => {
+            const start = performance.now()
+            setTimeout(() => runEffect(cancel("cancel_group"), middlewareContext), 250)
+            const effect = parallel(
+                interval(dispatch(TestActions.echo("1")), 500, { cancelId: "cancel_group" }),
+                timeout(dispatch(TestActions.echo("2")), 750, { cancelId: "cancel_group" }),
+                timeout(dispatch(TestActions.echo("3")), 1000, { cancelId: "cancel_group" })
+            )
+            await runEffect(effect, middlewareContext)
+            const end = performance.now()
+            await sleep(1000)
+
+            assert(end - start < 300, "Cancel group did not cancel all tasks")
+            assert(dispatched.length === 1)
+            assert(
+                Object.keys(middlewareContext.cancellables).length === 0,
+                "Cancel group did not clean up cancellable"
+            )
+        },
+        testRunningJobCancellationPrevention: async ({ assert, context: { logger, middlewareContext } }) => {
+            const effect = parallel(
+                timeout(
+                    async () => {
+                        await sleep(500)
+                        return dispatch(TestActions.echo(""))
+                    },
+                    10,
+                    { cancelId: "timeout" }
+                ),
+                timeout(cancel("timeout"), 100)
+            )
+            await runEffect(effect, middlewareContext)
+            assert(
+                middlewareContext.dispatched.length == 1,
+                "Cancel shouldn't cancel an already executing job in a timeout or interval"
             )
         },
         testTimeout: async ({ assert, fail, context: { middlewareContext, dispatched } }) => {
@@ -420,6 +462,32 @@ const Tests: TestGroup<typeof testContext> = {
                     store.getState().state2.echoReceived2 === true &&
                     logger.count() === 3 // all() from combine reducers and 2 dispatches
             )
+        },
+        testCombineRootEffectReducers: async ({ assert, context: { logger } }) => {
+            function rootEffectReducer1(state: State, action: TestActions) {
+                switch (action.type) {
+                    case "INCREMENT_NUMBER":
+                        return dispatch(TestActions.echo(`${state.state1.testNumber}`))
+                }
+            }
+            function rootEffectReducer2(state: State, action: Action) {
+                switch (action.type) {
+                    case "INCREMENT_NUMBER":
+                        return dispatch(TestActions.echo(`${state.state2.testNumber2}`))
+                }
+            }
+
+            const effectMiddleware = createEffectReducerMiddleware(
+                combineRootEffectReducers<State>(rootEffectReducer1, rootEffectReducer2),
+                { logger: logger }
+            )
+            const store = createStore(
+                combineReducers({ state1: reducer1, state2: reducer2 }),
+                applyMiddleware(effectMiddleware)
+            )
+
+            store.dispatch(TestActions.incrementNumber())
+            assert(logger.count() === 3 && isEqual(logger.last().effect, dispatch(TestActions.echo("5"))))
         },
         testEffectComparability: async ({ assert, fail }) => {
             for (const type of Object.values(EffectType)) {