+++ docs/release-notes.md
@@ -1,3 +1,12 @@
+# 1.2.1 (2020-05-01)
+## Added
+* Missing documentation link for `combineRootEffectReducers()`
+
+# 1.2.0 (2020-05-01)
+## Added
+* Cancel groups for reused cancel ids
+* `combineRootEffectReducers()` for multiple effect reducers that receive full state
+
 # 1.1.0 (2020-04-28)
 ## Added
 * Debounce Effect